Remove special casing for Any-annotated aliases
Summary:
> First, the z: Any situation looks like a bug or accidental feature to me.
This is definitely meant (and works) as a variable declaration; that it
also allows using z as a type seems wrong. I can't find any evidence in
PEP 484 that this was intended; in mypy it's likely the accidental result
of other design choices meant to shut up errors about Any.

Looks like `x: typing.Any` isn't actually intended to work as a valid type alias!

Remove support for this - if you want an alias to Any, you can declare as `x = typing.Any`.

This will hopefully open the doors to cleaning up some of our undefined vs undeclared vs any distinctions.

Reviewed By: mrkmndz

Differential Revision: D16692725

fbshipit-source-id: ac5a206b0afbb8b094eafbd63f6630dfc28d1c9f