Remove cyclic dependencies between `PythonTarget` and `BuildTargetsCollector`
Summary:
In previous series of refactoring, we finally get to the point where all other targets's build information are fully constructed before sending them to `BuildTargetsCollector`. This commit finishes the cleanup by finally remove the cyclic dependency between `PythonTarget` and `BuildTargetsCollector`.

To accomplish this, we need to pass `buckRoot` and `outputDirectory` to `PythonTarget.parse` so that it can construct absolute source mapping during construction time, so that we can simply put those source mappings directly into collect instead of constructing the absolute path at the time when we are about to add source mappings. This changes the content stored in `PythonTarget` so I also updated the tests.

Here is the beautiful picture after this cleanup:
- `XxxTarget` classes only know how to construct relevant build targets for builder.
- `BuildTargetsCollector` only collects build information. It delegates target specific json parsing to those `XxxTarget` classes.
- `BuildTargetsBuilder` only knows how to build things (already done).
- A much cleaner dependenecy chain: `BuildTargetsBuilder -> BuildTargetsCollector -> XxxTarget`.

Reviewed By: dkgi

Differential Revision: D16690824

fbshipit-source-id: 34847f3b3bfff6e352fe919b28d95485ee0e7d29