Break ((parsing + preprocessing) fixpoint) into (parsing + (preprocessing fixpoint))
Summary:
Previously, our fixpoint algorithm is written in such a way that if wildcard import resolution fails, we throw away everything that's been parsed and start from scratch again. This is wasteful. Let's make it such that import resolution failure would only restart the *preprocesser*, not the parser.

This is done by separating parsing and preprocessing into two different steps in `parse_sources`, where within each step we can split the work in parallel. The import resolution fixpoint can then be moved to and contained in the second step.

Reviewed By: dkgi

Differential Revision: D16702113

fbshipit-source-id: b72699be676f7fe168882c6b20e0e7402a4b193c