Purge `File.Handle` from LSP TextDocumentDefinitionResponse construction
Summary:
This is one of the two last remaining functionalities that depends on the old `File.Handle` search path construction algorithm instead of the new `ModuleTracker` one.

The strategy is similar to the diff below this: we do the full path expansion before calling into the message constructor. Expanding the full path needs to use un-instantiated `Reference.t` as key to lookup the corresponding `SourcePath.t`, but `LookupCache.find_definition` returns us already instantiated locations :(  Although we can "revert" this instantiation back with `SourcePath.qualifier_of_relative`, I figured that the more proper way of doing this is to simply return uninstantiated locations from `LookupCache.find_definition`. We should probably do this for other APIs in `LookupCache` as well, but that's not our immediate concern so let's put it into todo-list instead.

Reviewed By: dkgi

Differential Revision: D16714792

fbshipit-source-id: d2c70f6115ff78800a4749faa7c1523fdf71de23