Purge `File.Handle` from LSP publishDiagnostics message construction
Summary:
This is one of the two last remaining functionalities that depends on the old `File.Handle` search path construction algorithm instead of the new `ModuleTracker` one.

The situation here is a bit unfortunate: we only get relative paths from `TypeCheckResponse`, but we need full path for `PublishDiagnostics`. As a result, we can't just thread the path through and has to resort to this ugly `relative_to_full_path` workaround to bridge the gap :(  This can be done on the caller side so `PublishDIagnostics.of_errors` no longer need to worry about path translations.

Reviewed By: dkgi

Differential Revision: D16713576

fbshipit-source-id: 4643c3391f76bc5797573382cab755bb7d7ad252