Allowing for Concatenation in class parameterization
Summary:
The main motivator for introducing concatenation now was to enable experimentation with Tensors which have a unary parameter (a dtype) and a variadic parameter (for the lengths of the dimensions).
To do this, we convert our support for variadics in the class hierarchy edges to variadics, with leading and trailing unaries.
This was straightforwardly enabled by making our concatenation type generic (on the ocaml side), to allow for the representation of variadics flanked by unaries rather than mapped variadics flanked by arbitrary types.

Reviewed By: shannonzhu

Differential Revision: D16663193

fbshipit-source-id: 7ce2de0be2cc81f02c69163d14f56802b2c0cd34