Making the ClassHierarchy Handler interface read-only
Summary:
Now that we separated out the in memory handler only to tests, we can fold in the writing methods on the handler inside of `Analysis.Environment`, where they belong.  I then had to port all of the tests to interact with either the mock handler (typeorder and classhierarchy tests), which needed its own mutation methods, or with `Analysis.Environment` which is much cleaner and closer to prod :).

Note:

A lot of the line count here comes from having to re-order stuff in `Analysis.Environment`, which is kind of unfortunate re: diff bloat/ blame pollution, but seemed rather unavoidable

ohcaml

Reviewed By: grievejia

Differential Revision: D16713854

fbshipit-source-id: 01f3957711ab4d93d85a368988132401462c83f7