Simplifying Concatenation.Middle
Summary:
As promised in the discussion of D16597092, we can simplify the representation of `Concatenation.Middle`.

Rather than representing some non-zero numbers of mappers around a variable separately from an un-mapped variable, we instead can simply represent both as a combination of a variable and some (potentially zero) number of mappers.

This allows us to kill the OrderedTypes.Map module.

Reviewed By: shannonzhu, dkgi

Differential Revision: D16680415

fbshipit-source-id: 81231ed45831b5c4b6b45727056db3962768723b