merging Modules and ImplicitModule tables
Summary:
We can store the difference between these two things in the value, rather than in their table identity.
The fewer tables we have, the better.
Because we always read them together, this should be a no-op.  Also we are for some reason incrementing the count of implicit modules registered, but don't seem to be ever decrementing it.   Correspondingly we're never invalidating these... but that's a problem for later.

I also took this opportunity to inline the now simplified read methods.

Reviewed By: sinancepel

Differential Revision: D16724759

fbshipit-source-id: 3eff400835e5c6efb32fdcaa1635d1cb107fec8b