improve `pyre stop` experience by polling for pid
Summary:
`pyre stop` sends the command to stop the server on the OCaml side, but returns before verifying that the server process has had a chance to complete. There's an inherent race here, as the server process needs to send a message saying that the result was successful before terminating itself.

This polling is not necessary for correctness ( a pyre stop; sleep 1; pyre start would accomplish the same result, and we still kill if we haven't successfully polled the death of the process), but should improve the average user experience where `pyre restart` was failing with the confusing message of "Starting a Pyre server. A Pyre server already exists.".

Reviewed By: dkgi

Differential Revision: D16696300

fbshipit-source-id: 0dadc7b863530e04d666c3bc88df86ecd6b81ca8