Simplifying the interface of ClassHierarchy.Handler
Summary:
The interface of ClassHierarchy is unnecessarily complicated, when all it actually requires is a set of lookups which return optionals. All of the other operations (`exists`, `find_unsafe`) can trivially be implemented on the client side of this interface.
By simplifying this, we can finally (in the next diff), make the read-only interface of environment be a flat list of read-only tables, which makes dependency tracking much easier to think about and implement.

Reviewed By: grievejia

Differential Revision: D16742954

fbshipit-source-id: b12ebd06d13e4460b886e56c17c4cd3e1472fda9