Proof-of-concept of using hack-parallel dependency table
Summary:
Now that we have the interface of Analysis.Environment cleaned up, we can start working on how to represent the dependencies themselves.  If we were to try to store it into traditional shared memory, we run into the common problem of wanting to append from a worker.  We'd like to be able to look up the list of dependencies based on a given table/key pair, but adding these entries from the workers would require concurrent deletions/additions, which is not supported.  Instead, we can use the purpose built dependency table, which allows for concurrent appends to a given key, but with a limited key/value domain (32 bit integers each) and no deletion.  This means we can't use this table for a seemingly applicable use case where we need to do invalidation (i.e. backedges), but for dependency analysis, over-approximation is perfectly acceptable.

This diff cribs the interface code from hack, and writes up a proof of concept implementation of the hashing and decoding strategies that would be necessary for a more full-fledged implementation.

Next step will be to integrate this into the table creation functor to make a DependencyTrackedTable.

Reviewed By: grievejia

Differential Revision: D16752274

fbshipit-source-id: a14a525b0b7299e1351bfa17558c981bdb0d179d