Gracefully handle hash collisions in the dependency table
Summary:
Since we are hashing easily hundreds of thousands of dependencies, we are almost certainly going to have hash collisions in the dependency table.  To handle this, we approach this as a straightforward id assignment problem, where we use forward and backward lookup tables to uniquely assign ids to dependencies, without needing to do a reduce.

We only have to deal with this for the dependencies, not the table/key tuples, since the dependency table only goes one way (e.g. we are looking up the functions that depend on a class definition, not the other way around).  Thus we only need to decode those hashes.  Additionally, collisions in the table/key side are benign since that just over-approximates dependencies, while collisions in the dependency side would in the initial implementation lead to dropped dependencies.

Reviewed By: sinancepel

Differential Revision: D16775363

fbshipit-source-id: e738e6d59a58881a57b0773d968db8fb47351489