Fix type alias false positives on local variables
Summary:
This is a bit of a messy scoping issue -- we dequalify expressions before parsing them into types, so something like `$parameter$type` would turn into `Primitive type`. We'd flag this as an assignment to a type and then error that the type is missing type parameters.

We're somewhat saved here by the fact that local variables cannot be valid types.

```
def foo() -> None:
   x = int
   y: x = 1
```

will throw that `x` is an invalid type. So we can simply make sure we're only throwing these errors if the RHS is not local, in the case that it's a reference.

Reviewed By: MaggieMoss

Differential Revision: D16791098

fbshipit-source-id: 33482a4ef32614f180b5c58c023bed467494976e