Parameterize dependency table with dependency key type
Summary:
We were unnecessarily restricting the type of `dependency_value`. Furthermore, it's almost impossible to extend the datatype with any interesting type due to cyclic reference issue, as `memory.ml` was depended upon by most of our files.

Dependency values are themselves keys of other tables. Therefore it seems natural to piggyback on the existing `KeyType` signature, and require the type of the dependency values to be passed in as a module with signature `KeyType` in `DependencyTrackedTableWithCache` and `DependencyTrackedTableNoCache`. Dependency encoder can also be simplified by functorizing it w.r.t. `KeyType`: instead of building customized internal key module, we can simply take whatever `KeyType` module that gets passed in.

With this change, we are able to set `dependency_value` to any type we want, as long as we implement the corresponding `KeyType` interface.

Reviewed By: mrkmndz

Differential Revision: D16799604

fbshipit-source-id: 84153f892e2cd3f75b46e00f95291f04b806114c