Consider additional extensions in module tracker priority
Summary:
Avoid crashing the client when additional extensions are provided in the configuration, and module tracker gets confused choosing between files of the same name but different extensions.

We already have a prioritization system (.pyi over .pyi), ranking a.py vs a/__init__.py, and ranking based on which local root a file is found under.

If the existing prioritization can't make a decision, choose a module depending on the order of extensions provided in the configuration.

Note: if this diff lands, we should add this to the configuration documentation under the `extensions` field.

Reviewed By: grievejia

Differential Revision: D16806548

fbshipit-source-id: 038318aa70075a252f141c0d36bb3ec940f7a424