Fix server integration test
Summary:
It turns out our server integration test has been silently completely useless for a while now.
In `poor_mans_rsync`, it was blowing away the `.pyre` directory every time it moved up a commit, so we were always running from scratch, making the comparisons always succeed.

Furthermore, it was not exiting with a non-zero exit code when it ran out of retries, silencing whole other classes of issues (I don't think it's going through the proxy to github).

This test is failing now, correctly telling us that incremental for dependencies is not working at all.

Reviewed By: grievejia

Differential Revision: D16797270

fbshipit-source-id: 9c0fd38a1ac6f467ee8cf2abf38d766e1c1d988a