Enable transitive start
Summary: Transitive checking is only relevant on lsp events like file updated.  Therefore having that option on incremental does not make sense, and instead we need it on the start command in order to have that option saved for the future events.  We then can use that flag in the integration test, which doesn't yet fix it, but gets us closer.

Reviewed By: sinancepel

Differential Revision: D16818532

fbshipit-source-id: e38a27df3a1c9e1b0923a97383143f40bf3b7a0a