Don't crash on version checks
Summary:
Sandcastle rebase check jobs were sporadically failing, and it looks like this is because dewey sometimes fails to download typeshed when many version checks occur in parallel on a single machine (we schedule one for each project we are running pyre-check integration tests on).

Pyre-check doesn't run into this problem because we split each job onto a different machine - whereas for a quick version ping, this doesn't make sense to do. We can solve this issue by not crashing the client over typeshed validation checks when all that's needed is the version hash, and not construct a full configuration in this case. Just grab what's needed directly from the file instead.

Some examples:
https://our.intern.facebook.com/intern/sandcastle/job/18014398635243859
https://our.intern.facebook.com/intern/sandcastle/job/18014398635243426/
https://our.intern.facebook.com/intern/sandcastle/job/31525197519984075/

Reviewed By: grievejia

Differential Revision: D16821364

fbshipit-source-id: 75d443570d8af0ac9adca1e14842ab7c941c3eac