Support parsing `*type` in functions signature comment
Summary:
Mypy allows you to write type signature comments like `# type: (*str, **str) -> str` ([reference](https://mypy.readthedocs.io/en/latest/cheat_sheet.html#when-you-re-puzzled-or-when-things-are-complicated)). This is currently a parser error in Pyre since we don't even allow character '*' in the argument position. This commit adds the support for that.
It should be able to fix the parsing error in [#189](https://github.com/facebook/pyre-check/issues/189).

Reviewed By: dkgi

Differential Revision: D16832310

fbshipit-source-id: 32dc5dbd8d44e48264d1e3b225a1259ea70d0bc3