Error on using non-imported module
Summary:
We currently do not error on using non-imported module at all.
Minimum repro:
```
subprocess.check_output([])  # no Pyre errors, but there should be.
```

# High-level Approach:
1. Record what's imported in `Resolution.t`. We keep it in local environment so it can automatically handle imports in `if`, `def`, etc.
2. When type checking reference and we find that the reference is a module, check additionally whether it's properly imported.

# Temporary Solutions

## Type Query
Type query needs to have a way to getting the type of an expression without any imports. I choose to filter away using non-imported module error in type query to achieve that.

## `except (Exception1, Exception2):` handling

Some steps before type checking will turn that into `typing.Union[Exception1, Exception2]` and cause `typing.Union` to be type checked as a reference during type checking. This is a synthetic use of typing. To avoid the false positive, I think we can just ignore unused import of `typing` module for now, since changing the way we handle Try statement is more involved and should not be the focus of this diff.

Reviewed By: dkgi

Differential Revision: D16788798

fbshipit-source-id: 07a546e05bcc062451e3b478e631ea52324a6c3d