Track errors in liveness analysis state
Summary:
Spent some time thinking about the issue of scoping, dealing with variable definitions in branching states, etc.

The direction I think makes the most sense to go in is to keep identifiers that are still possible to use, i.e., in scope, in the map of unused identifiers. Once they become impossible to reference, we should convert them into errors.

This avoids the issue of having a list of locations associated with an identifier, and when that identifier is used, needing to figure out which of those locations have been used. This avoids janky heuristics and is also nice because it gets more complicated when you have a branching structure, so a single use might knock out several locations at once. Ex.

```
if check:
   x = 1
else:
   x = 2
use(x)
```

Next step is adding some scope information to unused identifiers to handle nested scopes. When we exit those scopes we'll do some error generation too.

Reviewed By: dkgi

Differential Revision: D16859091

fbshipit-source-id: 72388264f004317ef7e1ab0a1d57b78b3a73e874