Require value type of `DependencyTrackedTable` to be comparable
Summary:
This is purely an module interface change that does not affect any existing functionality (since nothing relies on `DependencyTrackedTable` yet).

Having a comparison interface will be useful in the followup diff, where we will introduce batch dependency computation. The usefulness comes from the fact that if the old and new values for the same key are compared equal, we don't need to add the key to the dependency set.

Reviewed By: dkgi

Differential Revision: D16841332

fbshipit-source-id: 286914c2e2c6f3af8b71db8494bb0f0306355ded