Compute `Ast.Source.hash` according to abstract syntax as opposed to concrete syntax
Summary:
I ran into this issue when trying to use the has value to determine whether the abstract syntax of a given `Source` is different from that of another `Source`, which is useful for the parser to filter out redundant incremental dependencies.

Apparently, `Source.hash` was previously computed based on the raw string of the input file. This is only used in saved state start, and once we have a better buck integration we might even be able to get rid of it. Figured that it is not what I want and its use case is rather limited, I rename it to `Source.Metadata.raw_hash` and write another `Source.hash` to replace the old one.

Reviewed By: sinancepel

Differential Revision: D16860729

fbshipit-source-id: 2ab97b70282628c5b5c1decb12dacdf1123a3f4b