Do not count AST dependency if there's no change in the abstract syntax
Summary: We can build this feature on top of the oldification/comparison dependency tracking mechanism introduced in D16851213 for  `DependencyTrackedTable`: Just promote the `Sources` table to a `DependencyTrackedTable`, and set each source to depend on itself. The `depracate` and `dependencies_since_last_deprecate` function will do their job for us. `Source.t` comparison is conducted based on `Source.hash` introduced in D16860729.

Reviewed By: dkgi

Differential Revision: D16878697

fbshipit-source-id: 7a8b0c6296bb0edc19bdb1755fd2fded49e4c914