Stop suppressing class toplevel errors
Summary:
Not sure why we were blanket suppressing all errors in class toplevels in default... class toplevels don't have a return annotation, and so we were considering them untyped functions and suppressing errors.

We had added a check to ensure the lack of a return value wouldn't suppress things in module toplevels but were missing such a check for class toplevels.

Reviewed By: grievejia

Differential Revision: D16907505

fbshipit-source-id: 8ffd99d2c0cd7a2b58c8449519c9bf1a49a8a42f