Handle wildcard import dependency in the parser
Summary:
Since wildcard exports are already stored in shared memory, tracking dependencies on it is not so different than what was done in D16878697:

Step1. Turn the `WildcardExports` table into a `DependencyTrackedTable`.
Step2. Insert the dependency correctly when expanding wildcard imports in the parser.
Step3. Deprecate and compute dependencies for the `WildcardExports` table along with the `Sources` table in `AstEnvironment.update_and_compute_dependencies`.

One more change I made here is that I try to sort the wildcard export list at creation time, so that diffing this list later can be easier.

Reviewed By: mrkmndz

Differential Revision: D16899549

fbshipit-source-id: 83d58d7e8311546096585ccd9400f786bc84c045