fix ignore handling for incremental checks
Summary:
We were registering ignores for all sources, not just the checked ones. Only make the Postprocess.ignore API public, and pass in checked_sources, so that we don't have to remember to call register_ignores on the right set of sources.

We didn't have a great way of testing incremental checks beforehand, so I introduced some scaffolding to make incremental check testing more human-readable - it uses ScratchProject under the hood and allows simulating "start with a set of X files, update to a complete state of Y files, and ensure we have the type errors at x, y, z."

Reviewed By: grievejia

Differential Revision: D16904775

fbshipit-source-id: 30c89997e4f7f1a7744d7a36ed2f0f5c4d152430