Split up purge
Summary:
The main thrust of this diff is separating the key lookup logic in purge from the actual act of deletion.
This is reasonable from an organizational perspective, but more importantly is a prerequisite for re-using the lookup logic in a dependency tracked update function, which will (under the hood) oldify those keys rather than delete them.

This is almost a NFC refactor, except for two small behavior alterations:
* Rather than setting edges(dead_keys) to [], it's instead just deleted.  This will dovetail nicely with the depedency tracked transaction workflow
* For the first time we actually do something to backedges(dead_keys), namely deleting them.  This should not have any functional impact since they are re-set to the empty set in SharedMemoryClassHierarchyHandler.insert, but is good for consistency's sake.

Reviewed By: sinancepel

Differential Revision: D16919875

fbshipit-source-id: ee21a43df978b998aeb41b8bbddf75bdb2476f70