Don't expose class keys table
Summary: It turns out that the class keys table is only used in calls that come from environment building anyway, so there is no real use in exposing it. On the other hand, exposing it as a dependency tracked table is problematic since every single call to it is to the same key, which ought to re-trigger everything, every time.

Reviewed By: dkgi

Differential Revision: D16925771

fbshipit-source-id: 6d16fcc9379cf973490cb477dc881ca0c794eff2