Propagate wildcard import
Summary:
# The problem

This is `a.py`:
```
from b import *
```
This is b.py:
```
from c import *
```
This is c.py:
```
derp = 42
```
Currently, `derp` exists in the wildcard export list of `c` and `b`, but not `a`: The export is propagated only one-level as opposed to all-the-way up.

# What this diff fixes

In end-to-end checks, `derp` will be in the wildcard export list of `c`, `b`, and `a`.

We would correctly identify that module `c` is a dependency of module `b`, and both module `b` and `c` are dependencies of module `a`

# What's not fixed

Incrementally updating `c.py` will not trigger a re-parse of file `b.py` and `a.py`.

We can't easily do the reparse because of how the current dependency tracking framework is written: it assumes that we structure the phases in a straight-line fashion: no dependency cycle is allowed among phases. Currently parsing, wildcard import resolution and the rest of the preprocessing passes are all put in the *same* phase, and unfortunately parsing+preprocessing depends on wildcard import resolution. At phase-level, this means that the parser phase depends on itself (i.e. a dependency cycle), which violates the basic assumption the framework requires.

We'll need a complete separation between the import resolution phase and the parsing&preprocessing phase to fully address this issue.

# How the fix works

Previously, we were only looking one-level up in `expand_wildcard_imports`. Let's change that to go all-the-way up.

Inevitably this is going to be work-inefficient: if worker0 processes a `x.py` that has `from z import *` and worker1 processes a file `y.py` that also has `from z import *`, transitive import resolution on module `z` will be computed by both workers. On the other hand, unless we want to maintain the dependencies by ourselves, we actually don't want the workers to share any works here, as it's going to mess up with the auto dependency tracking framework. For instance, if the content of `z.py` is actually `from w import *`, we'd want both worker0 and worker1 to know that `x` or `y` would transitively depend on `w` -- if we somehow make worker1 share the work done by worker0, we'd either lose the info that `y` depends on `w`, or we'd be forced to add all dependencies of `x` to `y`, potentially creating false dependencies.

In practice, the wasted work probably won't matter too much, as transitive wildcard import should be relatively rare.

Reviewed By: dkgi

Differential Revision: D16930605

fbshipit-source-id: b516fbfa32322c74d1ed58d6f3df4194792ecacc