Introducing new type checking dependency analysis behind a flag
Summary:
With the aid of dependency_tracked_purge_and_update and dependency_tracked_resolution, we can add the new kind of dependency analysis we've been trying to create.  This determines which files read certain shared memory tables * keys, and detects when those changes, then schedules those files for type checking.  Down the line the unit of type checking could likely be shrunk down to functions rather than files.

For testing purposes, this is ready, but not necessarily ready for prime time.  Because of that we protect this behind a flag.

Reviewed By: grievejia

Differential Revision: D16931084

fbshipit-source-id: 30582567dcd89ce8370e2a7a706b73124ca93e7c