Throw missing return even if state bottoms
Summary:
There was an issue reported last week where some functions without return annotations in strict mode did not error that the return was missing.

Turns out if we bottom out the state or if there is no body, we never throw the missing return - because we used to only ever throw this error when we ran into a return statement in the body.

Throw this error when checking the signature too, and that error will get joined with the others we throw at return statements if they exist.

Reviewed By: dkgi

Differential Revision: D17096617

fbshipit-source-id: f3a15bb73707bc05869ad33d207a42c011d72673