Refine type alias conditions
Summary:
We should only consider the RHS of an assignment as a type when it parses as a type and there is no annotation, or if the annotation is explicit type alias or meta.

May need to investigate whether meta is too general - currently it's necessary to handle TypedDicts and is at least a bit more specific than the check we had before.

Reviewed By: dkgi

Differential Revision: D17098311

fbshipit-source-id: be2fd006a86d326f4670990ccf6abe88bdc6f816