Make Annotation <-> Index global and immutable
Summary:
Id allocation is an intrinsically read + write kind of problem, as you need to validate that your id is free to claim it (mod concurrency strategies).  Therefore dependency tracking is problematic.  Furthermore, having the dependency keys of one table depend on the mutable values of another is confusing and fragile.

To avoid this, we can promote the association between a class name and its shorthand id to be a global immutable relationship, produced on-demand, avoiding the need for pre-registration and for invalidation.

Reviewed By: dkgi

Differential Revision: D17087726

fbshipit-source-id: 1eec3b27903a4d9604be3dc7aa6fcf918cbe3457