Add upgrade script support for switching to strict default
Summary:
Support the switch from default unsafe mode to default strict mode in the upgrade script.

Pass in a project, and the script will add the strict default to the configuration and add `# pyre-unsafe` to all files that need it (aka., were not marked as strict before and now generate errors in default strict mode).

Future todo: support removal of all `# pyre-strict` local headings which are now redundant with the default mode.

Reviewed By: dkgi

Differential Revision: D17101630

fbshipit-source-id: 5a7609448e9dc846d8cfd3ad52b901441915f879