Override variadic variable with target name
Summary: Because we don't correctly qualify the strings inside of the declaration of typevariables, we instead override their name with the target's name, at analysis/environment.ml:1078.  We were not doing the same thing for variadics, which was a problem when we wanted to do a round trip from type to expression and back, as when we generate implicit generic bases.

Reviewed By: grievejia

Differential Revision: D17136980

fbshipit-source-id: a390a4d56a58ac0e972375f8375f40d4ce169fb8