Connect implicit submodule tracking to prod
Summary:
Now that the new implicit submodule tracker is ready, let's connect it to prod!

`Environment.register_implicit_submodule` can now be thrown away. We expose `ModuleTracker.is_tracked_module` in `AstEnvironment` via `AstEnvironment.is_module`, and change `Environment.is_module` to simply forward the query to `AstEnviornment.is_module`.

As implicit submodules are no longer stored in the `Environment.Modules` table,  `GlobalResolution.module_definition` would be affected by this change as well, and we need to patch it up. To correctly return an implicit `Module.t` when an implicit module is looked up, `module_definition` needs to be able to tell implicit modules from explicit ones. Therefore we need to send the `AstEnvironment` to it as well.

Reviewed By: dkgi

Differential Revision: D17134129

fbshipit-source-id: 53cebb06af05e785955f6f8471901a497f7899ef