Track implicit submodules [1/2]
Summary:
We used to track implicit submodules [in shared memory](https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/fbcode/tools/pyre/analysis/environment.ml?commit=dac14ba786bfd1582ef59a2decfffa560737dd31&lines=932-945), which creates invalidation headaches for incremental analysis. The goal here is to move the tracking logic to `ModuleTracker`.

In our context, implicit modules are modules that do not have corresponding source file but exist anyway in Python runtime. For example, if we have file `a/b/c.py` under our local root, then
- `a.b.c` is an explicit module which has a corresponding source file `a/b/c.py`
- Both `a` and `a.b` are implicit modules -- their existence depends on the existence of `a.b.c`

The reason why we want to track these implicit modules is that our type checker depends on it to figure out whether a given module is a placeholder stub or not.

NOTE: Strictly speaking our notion of "implict module" does not fully capture what happens at Python runtime: in CPython interpreter they are not implicit and are actually backed up by filesystem directories: For example, if we create an empty directory `foo/` under local root, then module `foo` will exist at runtime but we don't count its existence in our implementation. Nevertheless, I think this model we have is still a reasonable approximation that serves as a good starting point.

The reason why we can't simply track these submodules in a plain hash set is because of invalidation. More specifically, according to our definition, different explicit modules may create the same implicit module. For example, if we have `a/b/c.py` and `a/b/d.py`, then both `a` and `a.b` are "shared" by the two explicit modules `a.b.c` and `a.b.d`. Say if we remove `a/b/d.py` in our incremental update -- we don't want to invalidate `a` and `a.b` immediately because they are still there due to the existence of `a/b/c.py`. But if we then remove `a/b/c.py` as well, we want `a` and `a.b` to be untracked from our system accordingly because we don't have any explicit modules that are responsible for their existence any more.

To handle incremental invalidation properly, we need to, for each module `M`, keep a counter to record how many module qualifiers have `M` as a prefix (including `M` itself). Turns out that simple ref-counting tricks would allow us to solve the invalidation issue. This diff only deals with setting the counters correctly upon `ModuleTracker` creation. In the next diff we are going to address how incremental updates are computed.

Reviewed By: dkgi

Differential Revision: D17075336

fbshipit-source-id: 468267dc95558241c2c1b3c788a418beaab1ceae