Converting typeshed_stubs to be put through the parsers
Summary:
Fundamentally it was weird that we were injecting typeshed after the parsing stage, as stand alone Source.t s.  This made it so our pre-environment couldn't parallelize on qualifiers since there were sources not in the ast environment.

Doing this conversion meant addressing more of the "__init__.py" problems.

Reviewed By: dkgi

Differential Revision: D17184093

fbshipit-source-id: c43c96721489eb45e8e82019b611c3ec5505fd99