derive comparison functions for type.ml
Summary: We were artisanally copying the functions needed for the Map and Set structures - it's possible to have the code generation take care of this for us. The additional benefit will be that when we use custom functions, it will catch the eye instead of getting lost in the boilerplate.

Reviewed By: dkgi

Differential Revision: D17190720

fbshipit-source-id: 20ab4de0f5ed61e0c7dffc7ed6d2341dd8101904