Standardize AST delta check as a decorator
Summary:
Responding again to suggestions in D17101630 to centralize the ast checking and make sure this happens any time we write code that can change file contents.

About to write another codemod to help deal with an upgrade containing lots of import errors that are a relatively simple fix - figured this should be implemented.

We're trading away some performance here (multiple file reads) for the simplicity of not having to break apart file-modifying functions into a piece that has already finished the parsing in order to feed into the decorator. Figured that kind of structure would defeat the point of having a modular check we can slap onto functions, but I'm not attached to this design and edits/suggestions are welcome.

Reviewed By: dkgi

Differential Revision: D17170438

fbshipit-source-id: 9d76e5c682b44258a1c005307ac1431543b6468a