Refactor codemods out of upgrade script
Summary:
I want to add another codemod to help with a new kind of error in the distillery upgrade and could potentially be used in fbcode projects with this issue too.

The upgrade script is where we put a lot of existing codemods that aren't actually upgrade-related, but it's starting to get cluttered and it felt wrong to keep adding to the huge pile of functions & helper functions for miscellaneous things.

Moving things around so upgrade.py contains only upgrade-specific functionality, and we can fool around in the codemod file with codemods that are useful for a bit and may be deleted later with less clutter.

Reviewed By: dkgi

Differential Revision: D17172624

fbshipit-source-id: d863fac1dac4f45c9b624107a015f5f633688ccb