Taking advantage of the UnannotatedGlobalEnvironment to parallelize and dependency track class definition registration
Summary:
With the class definition registration functionality successfully made into an independent phase, we can remove the existing serial implementation, and replace it with an update of the pre-environment.
Most of this diff is actually test maintenance, to port them over to this new registration system.  Furthermore because this approach iterates over qualifiers rather than sources, we have to continue our crusade against "`__init__.py`" sources, since again these clash with the builtins.pyi.

In the actual implementation, the flow of doing an update, and then using the new pre-environment is very straightforward, it's just buried under a mountain of test work.

Reviewed By: dkgi

Differential Revision: D17102370

fbshipit-source-id: a6369e02a829f157f1e91e97e1e0138f1df9ac3e